backend/task-worker/requirements.txt
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
python-telegram-bot==20.7
tweepy==4.14.0
facebook-sdk==3.1.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
stripe==7.6.0
sendgrid==6.10.0
twilio==8.10.0
slack-sdk==3.26.0
discord.py==2.3.2
prometheus-client==0.19.0 